Refactor patchWebAnimations into a new gulp task and use it everywhere it's needed
